docs: Update README.md
I spent unnecessary time trying to figure out how to correctly disable the `importReactOnTop` option.

To make it more intuitive for new users of this extension and to prevent similar confusion in the future, I suggest adding this explanatory line. It provides a clear example of how to reference the settings keys, making the configuration process smoother for everyone.